# Add support for CompletionList.Data.