--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSCompletionList.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSCompletionList.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
+using System.Linq;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
@@ -14,12 +15,19 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
 
         public Container<string> CommitCharacters { get; set; }
 
+        public object Data { get; set; }
+
         public static VSCompletionList Convert(CompletionList completionList, VSCompletionListCapability vsCompletionListCapability)
         {
             var vsCompletionList = new VSCompletionList(completionList);
             if (vsCompletionListCapability.CommitCharacters)
             {
                 PromoteCommonCommitCharactersOntoList(vsCompletionList);
+            }
+
+            if (vsCompletionListCapability.Data)
+            {
+                PromotedDataOntoList(vsCompletionList);
             }
 
             return vsCompletionList;
@@ -59,6 +67,24 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
                 {
                     completionItem.CommitCharacters = null;
                 }
+            }
+        }
+
+        private static void PromotedDataOntoList(VSCompletionList completionList)
+        {
+            // This piece makes a massive assumption that all completion items will have a resultId associated with them and their
+            // data properties will all be the same. Therefore, we can inspect the first item and empty out the rest.
+            var commonDataItem = completionList.FirstOrDefault();
+            if (commonDataItem == null)
+            {
+                // Empty list
+                return;
+            }
+
+            completionList.Data = commonDataItem.Data;
+            foreach (var completionItem in completionList.Items)
+            {
+                completionItem.Data = null;
             }
         }
     }