--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSCompletionListCapability.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/VSCompletionListCapability.cs
@@ -5,8 +5,16 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
 {
     internal class VSCompletionListCapability
     {
+        /// <summary>
+        /// Gets or sets a value indicating whether completion lists can have VSCommitCharacters. These commit characters get propagated
+        /// onto underlying valid completion items unless they have their own commit characters.
+        /// </summary>
         public bool CommitCharacters { get; set; }
 
+        /// <summary>
+        /// Gets or sets a value indicating whether completion lists can have Data bags. These data bags get propagated
+        /// onto underlying completion items unless they have their own data bags.
+        /// </summary>
         public bool Data { get; set; }
     }
 }