--- a/src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/VSCompletionListTest.cs
+++ b/src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/VSCompletionListTest.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Newtonsoft.Json.Linq;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using Xunit;
 
@@ -8,6 +9,60 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
 {
     public class VSCompletionListTest
     {
+        [Fact]
+        public void Convert_DataTrue_RemovesDataFromItems()
+        {
+            // Arrange
+            var dataObject = new JObject()
+            {
+                ["resultId"] = 123
+            };
+            var completionList = new CompletionList(
+                new CompletionItem()
+                {
+                    Label = "Test",
+                    Data = dataObject,
+                });
+            var capabilities = new VSCompletionListCapability()
+            {
+                Data = true,
+            };
+
+            // Act
+            var vsCompletionList = VSCompletionList.Convert(completionList, capabilities);
+
+            // Assert
+            Assert.Collection(vsCompletionList.Items, item => Assert.Null(item.Data));
+            Assert.Same(dataObject, vsCompletionList.Data);
+        }
+
+        [Fact]
+        public void Convert_DataFalse_DoesNotTouchData()
+        {
+            // Arrange
+            var dataObject = new JObject()
+            {
+                ["resultId"] = 123
+            };
+            var completionList = new CompletionList(
+                new CompletionItem()
+                {
+                    Label = "Test",
+                    Data = dataObject,
+                });
+            var capabilities = new VSCompletionListCapability()
+            {
+                Data = false,
+            };
+
+            // Act
+            var vsCompletionList = VSCompletionList.Convert(completionList, capabilities);
+
+            // Assert
+            Assert.Collection(vsCompletionList.Items, item => Assert.Same(dataObject, item.Data));
+            Assert.Null(vsCompletionList.Data);
+        }
+
         [Fact]
         public void Convert_CommitCharactersTrue_RemovesCommitCharactersFromItems()
         {