- `CompletionList.Data` serves as a way for completion items to not redundantly provide `Data`. Aka, all completion data objects serve as a default for underlying completion items.
- The LSP Platform hasn't yet inserted these changes into VS; however, our language server doesn't depend on the LSP platform directly anyhow. This changeset enables the data bag optimization of the LSP platform in a way that will automatically light up once it becomes available.
- Updated benchmarks to use new `Data` capability on the completion list.
- Updated our `VSCompletionListCapability` to include a `Data` property.
- Added tests

**Before:**
|                                              Method |     Mean |     Error |    StdDev |  Op/s |    Gen 0 |    Gen 1 |   Gen 2 | Allocated |
|---------------------------------------------------- |---------:|----------:|----------:|------:|---------:|---------:|--------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 7.434 ms | 0.1169 ms | 0.1036 ms | 134.5 | 585.9375 | 296.8750 |  7.8125 |  3,679 KB |
|           'Component Completion List Serialization' | 1.924 ms | 0.0257 ms | 0.0241 ms | 519.8 |  83.9844 |  37.1094 | 21.4844 |    523 KB |
|         'Component Completion List Deserialization' | 5.888 ms | 0.1090 ms | 0.1908 ms | 169.8 | 507.8125 | 242.1875 |       - |  3,156 KB |

**After:**
|                                              Method |       Mean |    Error |   StdDev |    Op/s |    Gen 0 |    Gen 1 | Gen 2 | Allocated |
|---------------------------------------------------- |-----------:|---------:|---------:|--------:|---------:|---------:|------:|----------:|
| 'Component Completion List Roundtrip Serialization' | 2,931.8 us | 43.66 us | 38.71 us |   341.1 | 304.6875 | 136.7188 |     - |  1,870 KB |
|           'Component Completion List Serialization' |   332.5 us |  6.04 us |  5.65 us | 3,007.8 |  26.8555 |   5.3711 |     - |    166 KB |
|         'Component Completion List Deserialization' | 2,836.9 us | 49.85 us | 46.63 us |   352.5 | 277.3438 | 121.0938 |     - |  1,704 KB |

**Results:**
Roundtrip: 2.5x faster
Serialization: 5.8x faster
Deserialization: 2.1x faster

Razor's language server part of dotnet/aspnetcore#31883